+++ crates/aiken-lang/src/parser/chain/field_access.rs
@@ -1,6 +1,6 @@
 use super::Chain;
 use crate::{
-    expr::UntypedExpr,
+    ast::well_known,
     parser::{token::Token, ParseError},
 };
 use chumsky::prelude::*;
@@ -8,25 +8,11 @@ use chumsky::prelude::*;
 pub(crate) fn parser() -> impl Parser<Token, Chain, Error = ParseError> {
     just(Token::Dot)
         .ignore_then(choice((
-            select! { Token::Else => "else".to_string() },
+            select! { Token::Else => well_known::VALIDATOR_ELSE.to_string() },
             select! { Token::Name { name } => name, },
+            select! { Token::UpName { name } => name, },
         )))
         .map_with_span(Chain::FieldAccess)
-}
-
-pub(crate) fn constructor() -> impl Parser<Token, UntypedExpr, Error = ParseError> {
-    select! {Token::Name { name } => name}
-        .map_with_span(|module, span| (module, span))
-        .then_ignore(just(Token::Dot))
-        .then(select! {Token::UpName { name } => name})
-        .map_with_span(|((module, m_span), name), span| UntypedExpr::FieldAccess {
-            location: span,
-            label: name,
-            container: Box::new(UntypedExpr::Var {
-                location: m_span,
-                name: module,
-            }),
-        })
 }
 
 #[cfg(test)]