+++ crates/aiken-lang/src/tests/check.rs
@@ -11,15 +11,23 @@ use crate::{
 };
 use std::collections::HashMap;
 
+const DEFAULT_MODULE_NAME: &str = "my_module";
+const DEFAULT_PACKAGE: &str = "test/project";
+
 fn parse(source_code: &str) -> UntypedModule {
+    parse_as(source_code, DEFAULT_MODULE_NAME)
+}
+
+fn parse_as(source_code: &str, name: &str) -> UntypedModule {
     let kind = ModuleKind::Lib;
-    let (ast, _) = parser::module(source_code, kind).expect("Failed to parse module");
+    let (mut ast, _) = parser::module(source_code, kind).expect("Failed to parse module");
+    ast.name = name.to_string();
     ast
 }
 
 fn check_module(
     ast: UntypedModule,
-    extra: Vec<(String, UntypedModule)>,
+    extra: Vec<UntypedModule>,
     kind: ModuleKind,
     tracing: Tracing,
 ) -> Result<(Vec<Warning>, TypedModule), (Vec<Warning>, Error)> {
@@ -31,27 +39,32 @@ fn check_module(
     module_types.insert("aiken".to_string(), builtins::prelude(&id_gen));
     module_types.insert("aiken/builtin".to_string(), builtins::plutus(&id_gen));
 
-    for (package, module) in extra {
+    for module in extra {
         let mut warnings = vec![];
+
+        if module.name == DEFAULT_MODULE_NAME {
+            panic!("passed extra modules with default name! Use 'parse_as' to define tests instead of 'parse'.");
+        }
 
         let typed_module = module
             .infer(
                 &id_gen,
                 kind,
-                &package,
+                DEFAULT_PACKAGE,
                 &module_types,
                 Tracing::All(TraceLevel::Verbose),
                 &mut warnings,
                 None,
             )
             .expect("extra dependency did not compile");
-        module_types.insert(package.clone(), typed_module.type_info.clone());
+
+        module_types.insert(typed_module.name.clone(), typed_module.type_info.clone());
     }
 
     let result = ast.infer(
         &id_gen,
         kind,
-        "test/project",
+        DEFAULT_PACKAGE,
         &module_types,
         tracing,
         &mut warnings,
@@ -76,7 +89,7 @@ fn check_with_verbosity(
 
 fn check_with_deps(
     ast: UntypedModule,
-    extra: Vec<(String, UntypedModule)>,
+    extra: Vec<UntypedModule>,
 ) -> Result<(Vec<Warning>, TypedModule), (Vec<Warning>, Error)> {
     check_module(ast, extra, ModuleKind::Lib, Tracing::verbose())
 }
@@ -2255,6 +2268,524 @@ fn allow_expect_into_opaque_type_constructor_without_typecasting_in_module() {
 }
 
 #[test]
+fn use_imported_type_as_namespace_for_patterns() {
+    let dependency = r#"
+        pub type Credential {
+          VerificationKey(ByteArray)
+          Script(ByteArray)
+        }
+    "#;
+
+    let source_code = r#"
+        use cardano/address.{Credential, Script, VerificationKey}
+
+        pub fn compare(left: Credential, right: Credential) -> Ordering {
+          when left is {
+            Script(left) ->
+              when right is {
+                Script(right) -> Equal
+                _ -> Less
+              }
+            VerificationKey(left) ->
+              when right is {
+                Script(_) -> Greater
+                VerificationKey(right) -> Equal
+              }
+          }
+        }
+    "#;
+
+    let result = check_with_deps(
+        parse(source_code),
+        vec![(parse_as(dependency, "cardano/address"))],
+    );
+
+    assert!(matches!(result, Ok(..)), "{result:#?}");
+}
+
+#[test]
+fn use_type_as_namespace_for_patterns() {
+    let dependency = r#"
+        pub type Foo {
+          A { a: Int }
+          B
+        }
+    "#;
+
+    let source_code = r#"
+        use thing.{Foo}
+
+        fn bar(foo: Foo) {
+          when foo is {
+            Foo.A { a } -> a > 10
+            Foo.B -> False
+          }
+        }
+    "#;
+
+    let result = check_with_deps(parse(source_code), vec![(parse_as(dependency, "thing"))]);
+
+    assert!(matches!(result, Ok(..)), "{result:#?}");
+}
+
+#[test]
+fn use_nested_type_as_namespace_for_patterns() {
+    let dependency = r#"
+        pub type Foo {
+          A { a: Int }
+          B
+        }
+    "#;
+
+    let source_code = r#"
+        use foo.{Foo}
+
+        fn bar(x: Foo) {
+          when x is {
+            foo.Foo.A { a } -> a > 10
+            foo.Foo.B -> False
+          }
+        }
+    "#;
+
+    let result = check_with_deps(parse(source_code), vec![(parse_as(dependency, "foo"))]);
+
+    assert!(matches!(result, Ok(..)), "{result:#?}");
+}
+
+#[test]
+fn use_opaque_type_as_namespace_for_patterns_fails() {
+    let dependency = r#"
+        pub opaque type Foo {
+          A { a: Int }
+          B
+        }
+    "#;
+
+    let source_code = r#"
+        use thing.{Foo}
+
+        fn bar(foo: Foo) {
+          when foo is {
+            Foo.A { a } -> a > 10
+            Foo.B -> False
+          }
+        }
+    "#;
+
+    let result = check_with_deps(parse(source_code), vec![(parse_as(dependency, "thing"))]);
+
+    assert!(
+        matches!(
+            &result,
+            Err((warnings, Error::UnknownTypeConstructor { name, .. })) if name == "A" && warnings.is_empty(),
+        ),
+        "{result:#?}"
+    );
+}
+
+#[test]
+fn use_wrong_type_as_namespace_for_patterns_fails() {
+    let dependency = r#"
+        pub type Foo {
+          A { a: Int }
+          B
+        }
+
+        pub type Bar {
+          C
+          D
+        }
+    "#;
+
+    let source_code = r#"
+        use thing.{Foo}
+
+        fn bar(foo: Foo) {
+          when foo is {
+            Foo.A { .. } -> True
+            Foo.D -> False
+          }
+        }
+    "#;
+
+    let result = check_with_deps(parse(source_code), vec![(parse_as(dependency, "thing"))]);
+
+    assert!(
+        matches!(
+            &result,
+            Err((warnings, Error::UnknownTypeConstructor { name, .. })) if name == "D" && warnings.is_empty(),
+        ),
+        "{result:#?}"
+    );
+}
+
+#[test]
+fn use_type_as_namespace_for_constructors() {
+    let dependency = r#"
+        pub type Foo {
+          I { i: Int }
+          B(Bool)
+        }
+    "#;
+
+    let source_code = r#"
+        use foo.{Foo}
+
+        test my_test() {
+          and {
+            Foo.I { i: 42 } == foo.I(42),
+            foo.B(True) == Foo.B(True),
+          }
+        }
+    "#;
+
+    let result = check_with_deps(parse(source_code), vec![(parse_as(dependency, "foo"))]);
+
+    assert!(matches!(result, Ok(..)), "{result:#?}");
+}
+
+#[test]
+fn use_type_as_nested_namespace_for_constructors() {
+    let dependency = r#"
+        pub type Foo {
+          I { i: Int }
+          B(Bool)
+        }
+    "#;
+
+    let source_code = r#"
+        use foo
+
+        test my_test() {
+          trace foo.Foo.I { i: 42 }
+          trace foo.Foo.B(False)
+          Void
+        }
+    "#;
+
+    let result = check_with_deps(parse(source_code), vec![(parse_as(dependency, "foo"))]);
+
+    assert!(matches!(result, Ok(..)), "{result:#?}");
+}
+
+#[test]
+fn use_type_as_nested_namespace_for_constructors_from_multi_level_module() {
+    let dependency = r#"
+        pub type Foo {
+          I { i: Int }
+          B(Bool)
+        }
+    "#;
+
+    let source_code = r#"
+        use foo/bar
+
+        test my_test() {
+          trace bar.Foo.I { i: 42 }
+          trace bar.Foo.B(True)
+          Void
+        }
+    "#;
+
+    let result = check_with_deps(parse(source_code), vec![(parse_as(dependency, "foo/bar"))]);
+
+    assert!(matches!(result, Ok(..)), "{result:#?}");
+}
+
+#[test]
+fn use_type_as_nested_namespace_for_constructors_from_module_alias() {
+    let dependency = r#"
+        pub type Foo {
+          I { i: Int }
+          B(Bool)
+        }
+    "#;
+
+    let source_code = r#"
+        use foo as bar
+
+        test my_test() {
+          trace bar.Foo.I { i: 42 }
+          trace bar.Foo.B(True)
+          Void
+        }
+    "#;
+
+    let result = check_with_deps(parse(source_code), vec![(parse_as(dependency, "foo"))]);
+
+    assert!(matches!(result, Ok(..)), "{result:#?}");
+}
+
+#[test]
+fn use_type_as_namespace_unknown_constructor() {
+    let dependency = r#"
+        pub type Foo {
+          I { i: Int }
+          B(Bool)
+        }
+    "#;
+
+    let source_code = r#"
+        use foo.{Foo}
+
+        test my_test() {
+          Foo.I(42) == Foo.A
+        }
+    "#;
+
+    let result = check_with_deps(parse(source_code), vec![(parse_as(dependency, "foo"))]);
+
+    assert!(
+        matches!(
+            &result,
+            Err((warnings, Error::UnknownTypeConstructor { name, .. })) if name == "A" && warnings.is_empty(),
+        ),
+        "{result:#?}"
+    );
+}
+
+#[test]
+fn use_wrong_type_as_namespace() {
+    let dependency = r#"
+        pub type Foo {
+          I { i: Int }
+          B(Bool)
+        }
+
+        pub type Bar {
+          S { s: String }
+          L(List<Int>)
+        }
+    "#;
+
+    let source_code = r#"
+        use foo.{Foo}
+
+        test my_test() {
+          trace Foo.S(@"wut")
+          Void
+        }
+    "#;
+
+    let result = check_with_deps(parse(source_code), vec![(parse_as(dependency, "foo"))]);
+
+    assert!(
+        matches!(
+            &result,
+            Err((warnings, Error::UnknownTypeConstructor { name, .. })) if name == "S" && warnings.is_empty(),
+        ),
+        "{result:#?}"
+    );
+}
+
+#[test]
+fn use_wrong_nested_type_as_namespace() {
+    let dependency = r#"
+        pub type Foo {
+          I { i: Int }
+          B(Bool)
+        }
+
+        pub type Bar {
+          S { s: String }
+          L(List<Int>)
+        }
+    "#;
+
+    let source_code = r#"
+        use foo
+
+        test my_test() {
+          trace foo.Foo.S(@"wut")
+          Void
+        }
+    "#;
+
+    let result = check_with_deps(parse(source_code), vec![(parse_as(dependency, "foo"))]);
+
+    assert!(
+        matches!(
+            &result,
+            Err((warnings, Error::UnknownTypeConstructor { name, .. })) if name == "S" && warnings.is_empty(),
+        ),
+        "{result:#?}"
+    );
+}
+
+#[test]
+fn use_private_type_as_nested_namespace_fails() {
+    let dependency = r#"
+        type Foo {
+          I { i: Int }
+          B(Bool)
+        }
+    "#;
+
+    let source_code = r#"
+        use foo
+
+        test my_test() {
+          trace foo.Foo.I { i: 42 }
+          Void
+        }
+    "#;
+
+    let result = check_with_deps(parse(source_code), vec![(parse_as(dependency, "foo"))]);
+
+    assert!(
+        matches!(
+            &result,
+            Err((warnings, Error::UnknownModuleType { name, .. })) if name == "Foo" && warnings.is_empty(),
+        ),
+        "{result:#?}"
+    );
+}
+
+#[test]
+fn use_opaque_type_as_namespace_for_constructors_fails() {
+    let dependency = r#"
+        pub opaque type Foo {
+          I { i: Int }
+          B(Bool)
+        }
+    "#;
+
+    let source_code = r#"
+        use foo.{Foo}
+
+        test my_test() {
+          trace Foo.I(42)
+          Void
+        }
+    "#;
+
+    let result = check_with_deps(parse(source_code), vec![(parse_as(dependency, "foo"))]);
+
+    assert!(
+        matches!(
+            &result,
+            Err((warnings, Error::UnknownTypeConstructor { name, .. })) if name == "I" && warnings.is_empty(),
+        ),
+        "{result:#?}"
+    );
+}
+
+#[test]
+fn use_opaque_type_as_nested_namespace_for_constructors_fails() {
+    let dependency = r#"
+        pub opaque type Foo {
+          I { i: Int }
+          B(Bool)
+        }
+    "#;
+
+    let source_code = r#"
+        use foo
+
+        test my_test() {
+          trace foo.Foo.I(42)
+          Void
+        }
+    "#;
+
+    let result = check_with_deps(parse(source_code), vec![(parse_as(dependency, "foo"))]);
+
+    assert!(
+        matches!(
+            &result,
+            Err((warnings, Error::UnknownTypeConstructor { name, .. })) if name == "I" && warnings.is_empty(),
+        ),
+        "{result:#?}"
+    );
+}
+
+#[test]
+fn use_non_imported_module_as_namespace() {
+    let dependency = r#"
+        pub type Foo {
+          I(Int)
+          B(Bool)
+        }
+    "#;
+
+    let source_code = r#"
+        test my_test() {
+          trace foo.Foo.I(14)
+          Void
+        }
+    "#;
+
+    let result = check_with_deps(parse(source_code), vec![(parse_as(dependency, "foo"))]);
+
+    assert!(
+        matches!(
+            &result,
+            Err((warnings, Error::UnknownModule { name, .. })) if name == "foo" && warnings.is_empty(),
+        ),
+        "{result:#?}"
+    );
+}
+
+#[test]
+fn invalid_type_field_access_chain() {
+    let dependency = r#"
+        pub type Foo {
+          I(Int)
+          B(Bool)
+        }
+    "#;
+
+    let source_code = r#"
+        use foo.{Foo}
+
+        test my_test() {
+          trace Foo.I.Int(42)
+          Void
+        }
+    "#;
+
+    let result = check_with_deps(parse(source_code), vec![(parse_as(dependency, "foo"))]);
+
+    assert!(
+        matches!(
+            &result,
+            Err((warnings, Error::InvalidFieldAccess { .. })) if warnings.is_empty(),
+        ),
+        "{result:#?}"
+    );
+}
+
+#[test]
+fn invalid_type_field_access_chain_2() {
+    let dependency = r#"
+        pub type Foo {
+          I(Int)
+          B(Bool)
+        }
+    "#;
+
+    let source_code = r#"
+        use foo.{Foo}
+
+        test my_test() {
+          trace Foo.i(42)
+          Void
+        }
+    "#;
+
+    let result = check_with_deps(parse(source_code), vec![(parse_as(dependency, "foo"))]);
+
+    assert!(
+        matches!(
+            &result,
+            Err((warnings, Error::UnknownTypeConstructor { .. })) if warnings.is_empty(),
+        ),
+        "{result:#?}"
+    );
+}
+
+#[test]
 fn forbid_importing_or_using_opaque_constructors() {
     let dependency = r#"
         pub opaque type Thing {
@@ -2264,7 +2795,7 @@ fn forbid_importing_or_using_opaque_constructors() {
     "#;
 
     let source_code = r#"
-        use foo/thing.{Thing, Foo}
+        use thing.{Thing, Foo}
 
         fn bar(thing: Thing) {
           expect Foo(a) = thing
@@ -2273,15 +2804,12 @@ fn forbid_importing_or_using_opaque_constructors() {
     "#;
 
     assert!(matches!(
-        check_with_deps(
-            parse(source_code),
-            vec![("foo/thing".to_string(), parse(dependency))],
-        ),
+        check_with_deps(parse(source_code), vec![(parse_as(dependency, "thing"))],),
         Err((_, Error::UnknownModuleField { .. })),
     ));
 
     let source_code = r#"
-        use foo/thing.{Thing}
+        use thing.{Thing}
 
         fn bar(thing: Thing) {
           expect Foo(a) = thing
@@ -2290,10 +2818,7 @@ fn forbid_importing_or_using_opaque_constructors() {
     "#;
 
     assert!(matches!(
-        check_with_deps(
-            parse(source_code),
-            vec![("foo/thing".to_string(), parse(dependency))],
-        ),
+        check_with_deps(parse(source_code), vec![(parse_as(dependency, "thing"))],),
         Err((_, Error::UnknownTypeConstructor { .. })),
     ));
 }